+++ webdriver-precore/rio-poc/src/WebDriver/RIO/App.hs
@@ -5,21 +5,22 @@
 -- Provides a runner function to set up a BaseEnv with logging and
 -- capabilities, then execute a RIO action in that context.
 module WebDriver.RIO.App
-  ( runner,
+  ( runWebDriver,
   )
 where
 
 import RIO
 import WebDriver.RIO.Capabilities (FullCapabilities)
 import WebDriver.RIO.Env (BaseEnv (..))
+import WebDriver.RIO.Logging (LoggerConfig, withLogging)
 
 -- | Initialize a BaseEnv context and run a RIO action.
 --
--- Takes log options and capabilities, creates a BaseEnv, logs a debug
+-- Takes logger config and capabilities, creates a BaseEnv, logs a debug
 -- message, then executes the provided action.
-runner :: LogOptions -> FullCapabilities cap -> RIO (BaseEnv cap) a -> IO a
-runner logOpts caps action =
-  withLogFunc logOpts $ \lf -> do
+runWebDriver :: LoggerConfig -> FullCapabilities cap -> RIO (BaseEnv cap) a -> IO a
+runWebDriver loggerConfig caps action =
+  withLogging loggerConfig $ \lf -> do
     let env = MkBaseEnv lf caps
     runRIO env $ do
       logDebug "Running action"