+++ webdriver-precore/test-utils/src/WebDriverPreCore/Test/CapabilitiesBuilder.hs
@@ -1,4 +1,4 @@
-module WebDriverPreCore.Test.RuntimeConst
+module WebDriverPreCore.Test.CapabilitiesBuilder
   ( httpCapabilities,
     httpFullCapabilities,
   )